FEAT: Add support for continiuous phenotypes
Bitepi only works over case/control phenotypes, but there is demand for
it to also accept input where the phenotype is a float.

Rework the Epistat and Dataset classes to be templated on whether the
input phenotype is continuous or not. The continuous behaviour can be
triggered with the command line argument `-cont`. It uses a measure of
variance that is identical to GINI for float values of 0 and 1, namely
the beta value is calculated as 1 - 2*variance. The variance calculation
is much slower than the simple counting in the case/control mode, on the
order of 4x slower for 20SNPs/1000samples, and 8x slower for
200SNPs/10000samples.