REFACTOR: Calculate variance in a single pass
Calculating the mean and then the variance requires storing all the
floats in the contingency table. This, combined with then going through
each contingency cell's array of floats means the run time of the
continuous phenotype version is much slower than the case/control
method.

Switch to a single pass variance calculation method, so the floats don't
need to be repeatedly stored after the initial read. This reduces the
run time to around 2x the case/control version. Note that floats have
been increased to doubles to reduce the chance of a catastrophic
cancellation when subtracting the squared sum from the sum of squares.
Alternatively an algorithm like Welford's algorithm could be used, which
won't be affected by such errors but takes around twice as long.